Correction modifySauce du controlleur sauce pour les modifications d'images

+++ back/controllers/sauces.js
@@ -93,24 +93,53 @@ exports.deleteSauce = (req, res, next) => {
 // Modification d'une sauce de la BDD par son créateur
 
 exports.modifySauce = (req, res, next) => {
-    Sauce.findOne({ _id: req.params.id }).then((sauce) => {
-        const filename = sauce.imageUrl.split("/images/")[1];
-        fs.unlink(`images/${filename}`, () => {
-            const sauceObject = req.file
-                ? {
-                    ...JSON.parse(req.body.sauce),
-                    imageUrl: `${req.protocol}://${req.get("host")}/images/${req.file.filename
-                        }`,
-                }
-                : { ...req.body };
-            Sauce.updateOne(
-                { _id: req.params.id },
-                { ...sauceObject, _id: req.params.id }
-            )
-                .then(() => res.status(200).json({ message: "Votre sauce a bien été modifiée !" }))
-                .catch((error) => res.status(400).json({ error }));
-        });
-    });
+
+    // Si la modification de la sauce concerne l'image 
+
+    if (req.file) {
+
+        // On identifie la sauce par son ID unique
+
+        Sauce.findOne({ _id: req.params.id })
+            .then(sauce => {
+
+                // On supprime son image et le lien entre l'ancienne image et la sauce concernée
+
+                const filename = sauce.imageUrl.split("/images/")[1];
+                fs.unlink(`images/${filename}`, () => {
+
+                    // On update les infos de la sauce et sa nouvelle image
+
+                    const sauceObject = {
+                        ...JSON.parse(req.body.sauce),
+                        imageUrl: `${req.protocol}://${req.get("host")}/images/${req.file.filename}`,
+                    };
+
+                    Sauce.updateOne(
+                        { _id: req.params.id },
+                        { ...sauceObject, _id: req.params.id }
+                    )
+                        .then(() => res.status(200).json({ message: "Votre sauce a bien été modifiée !" }))
+                        .catch(error => res.status(400).json({ error }));
+                });
+            })
+            .catch(error => res.status(500).json({ error }));
+    }
+
+    // Si la modification de la sauce ne concerne pas l'image
+
+    else {
+
+        // On update les infos de la sauce concernée
+
+        const sauceObject = { ...req.body };
+        Sauce.updateOne(
+            { _id: req.params.id },
+            { ...sauceObject, _id: req.params.id }
+        )
+            .then(() => res.status(200).json({ message: "Votre sauce a bien été modifiée !" }))
+            .catch(error => res.status(400).json({ error }));
+    }
 };
 
 // Gestion des avis Like et Dislike des sauces
